moving the new constants into the updated checklist
